+++ gym_space_engineers/__init__.py
@@ -34,6 +34,17 @@ register(
 )
 
 register(
+    id="SE-Forward-v1",
+    entry_point="gym_space_engineers.envs:WalkingRobotIKEnv",
+    max_episode_steps=200,  # around 20s of interaction
+    kwargs={
+        "control_frequency": 10,  # 10Hz
+        "symmetric_control": False,
+        "add_end_effector_velocity": True,
+    },
+)
+
+register(
     id="SE-Symmetric-v1",
     entry_point="gym_space_engineers.envs:WalkingRobotIKEnv",
     max_episode_steps=200,  # around 20s of interaction
@@ -41,7 +52,6 @@ register(
         "control_frequency": 10,  # 10Hz
         "symmetric_control": True,
         "add_end_effector_velocity": True,
-        # "allowed_leg_angle": 25, # for small slopes, 15deg by default
     },
 )
 
@@ -51,7 +61,7 @@ register(
     max_episode_steps=200,  # around 20s of interaction
     kwargs={
         "control_frequency": 10,  # 10Hz
-        "symmetric_control": True,
+        "symmetric_control": False,
         "add_end_effector_velocity": True,
         "task": "turn_left",
     },
@@ -80,7 +90,7 @@ register(
     max_episode_steps=200,  # around 20s of interaction
     kwargs={
         "control_frequency": 10,  # 10Hz
-        "symmetric_control": True,
+        "symmetric_control": False,
         "add_end_effector_velocity": True,
         "task": "forward",
         "randomize_task": True,